requirements.txt
apscheduler
Django>=3.1.8,<3.2
django-apscheduler
django-cors-headers
django-extensions
django-filter
django-health-check
django-import-export
djangorestframework
markdown
pandas
plotly >= 4.14
requests-toolbelt
